Organize canister method tests into common folder
