+++ property_tests/tests/canister_methods/init/dfx.json
