+++ property_tests/tests/canister_methods/init/package.json
