+++ property_tests/tests/canister_methods/init/test/generate_callable_method_body.ts
