+++ property_tests/tests/canister_methods/init/test/generate_init_body.ts
