+++ property_tests/tests/canister_methods/init/tsconfig.json
