+++ property_tests/tests/canister_methods/post_upgrade/package.json
