+++ property_tests/tests/canister_methods/post_upgrade/test/generate_callable_method_body.ts
