+++ property_tests/tests/canister_methods/post_upgrade/test/generate_init_method_body.ts
