+++ property_tests/tests/canister_methods/post_upgrade/test/generate_init_tests.ts
@@ -17,7 +17,9 @@ export function generateTests(
             {
                 name: `init method`,
                 test: async () => {
-                    const actor = getActor('./tests/post_upgrade_method/test');
+                    const actor = getActor(
+                        './tests/canister_methods/post_upgrade/test'
+                    );
 
                     const initValues = await actor.getInitValues();
                     const isPostUpgradeCalled =
