+++ property_tests/tests/canister_methods/post_upgrade/test/global_var_name.ts
