+++ property_tests/tests/canister_methods/query/dfx.json
