+++ property_tests/tests/canister_methods/query/package.json
