+++ property_tests/tests/canister_methods/query/test/generate_body.ts
