+++ property_tests/tests/canister_methods/query/tsconfig.json
