+++ property_tests/tests/canister_methods/update/package.json
