+++ property_tests/tests/canister_methods/update/test/generate_body.ts
