+++ property_tests/tests/canister_methods/update/tsconfig.json
